+++ src/control/sessions.rs
@@ -1,11 +1,11 @@
 use std::fmt::Debug;
 
-use axum::{Extension, Json, http::StatusCode};
+use axum::{Extension, Json, http::StatusCode, response::{IntoResponse, AppendHeaders}};
 use axum_extra::extract::{CookieJar, cookie::Cookie};
 use cookie::time::OffsetDateTime;
-use tracing::info;
+use tracing::{info, warn};
 
-use crate::{domain::users::Credentials, service::sessions::{SessionService, LoginError}, constants::SESSION_COOKIE_NAME};
+use crate::{domain::users::Credentials, service::sessions::{SessionService, LoginError, SessionVerifyError}, constants::{SESSION_COOKIE_NAME, USER_ID_HEADER}};
 
 #[tracing::instrument(skip_all, fields(email = credentials.email))]
 pub async fn post_sessions<T: SessionService + Debug>(
@@ -31,4 +31,35 @@ pub async fn post_sessions<T: SessionService + Debug>(
         .finish();
 
     Ok((jar.add(cookie), StatusCode::CREATED))
+}
+
+#[tracing::instrument(skip_all)]
+pub async fn get_sessions<T: SessionService + Debug>(
+    Extension(service): Extension<T>,
+    jar: CookieJar
+) -> Result<impl IntoResponse, StatusCode> {
+    info!("Received login attempt");
+    let session_id = match jar.get(SESSION_COOKIE_NAME.as_str()) {
+        Some(cookie) => cookie.value(),
+        None => {
+            warn!("No session ID provided!");
+            return Err(StatusCode::UNAUTHORIZED);
+        }
+    };
+
+    let user = match service.verify(session_id).await {
+        Err(SessionVerifyError::Missing) =>  {
+            return Err(StatusCode::UNAUTHORIZED);
+        },
+        Err(SessionVerifyError::Unknown) => {
+            return Err(StatusCode::INTERNAL_SERVER_ERROR);
+        },
+        Ok(session) => session
+    };
+
+    let headers = AppendHeaders([
+        (USER_ID_HEADER.as_str(), user.id)
+    ]);
+
+    Ok(headers)
 }
