+++ src/routing/mod.rs
@@ -1,10 +1,11 @@
+mod sessions;
 mod users;
 
-use axum::{Router, Extension};
+use axum::Router;
 
-use crate::{control::sessions::post_sessions, service::{sessions::HashSessionService, hash::BcryptHashService, users::HashUserService}, repository::{sessions::HttpSessionRepository, users::HttpUserRepository}, constants::{RESOURCE_MANAGEMENT_URL, SESSION_ID_GEN_RETRIES}};
+use crate::{service::{sessions::HashSessionService, hash::BcryptHashService, users::HashUserService}, repository::{sessions::HttpSessionRepository, users::HttpUserRepository}, constants::{RESOURCE_MANAGEMENT_URL, SESSION_ID_GEN_RETRIES}};
 
-use self::users::users_router;
+use self::{users::users_router, sessions::sessions_router};
 
 pub fn main_router() -> Router {
     let users_url = RESOURCE_MANAGEMENT_URL.clone() + "/users";
@@ -21,10 +22,7 @@ pub fn main_router() -> Router {
         *SESSION_ID_GEN_RETRIES
     );
 
-    let sessions_handler = axum::routing::post(post_sessions::<HashSessionService<HttpSessionRepository, HttpUserRepository, BcryptHashService>>);
-
     Router::new()
         .nest("/users", users_router(users_service))
-        .route("/sessions", sessions_handler)
-        .layer(Extension(sessions_service))
+        .nest("/sessions", sessions_router(sessions_service))
 }
